Replace test file to main
